fix: eliminate MSVC compiler warnings and upgrade Jsoncpp API
Fixed all C4996 deprecation warnings and C4244 type conversion warnings:

1. Jsoncpp Deprecated API Migration:
   - Replaced Json::Reader with Json::CharReaderBuilder + parseFromStream()
   - Replaced Json::FastWriter with Json::StreamWriterBuilder + writeString()
   - Replaced Json::StyledWriter with Json::StreamWriterBuilder + writeString()
   - Added helper functions for consistent JSON parsing/serialization

2. Type Conversion Fixes:
   - Changed TTL variables from 'long' to 'long long' in OAuth2Plugin.h
   - Fixed time-to-live calculations in CachedOAuth2Storage.cc
   - Resolved 64-bit Windows data truncation issues (long is 4 bytes)

3. Compiler Warning Suppression:
   - Added _SILENCE_CXX17_CODECVT_HEADER_DEPRECATION_WARNING for codecvt
   - Added _CRT_SECURE_NO_WARNINGS for getenv() function
   - Applied only to MSVC builds using conditional compilation

4. Cross-Platform Compatibility:
   - All changes use standard C++ types and APIs
   - Jsoncpp APIs are cross-platform compatible
   - Platform-specific code isolated in MSVC conditionals

Modified Files:
- OAuth2Backend/CMakeLists.txt: Added MSVC warning suppression macros
- OAuth2Backend/test/CMakeLists.txt: Added MSVC warning suppression macros
- OAuth2Backend/plugins/OAuth2Plugin.h: Fixed type conversion warnings
- OAuth2Backend/storage/RedisOAuth2Storage.cc: Upgraded Jsoncpp API
- OAuth2Backend/storage/CachedOAuth2Storage.cc: Upgraded Jsoncpp API + type fixes
- OAuth2Backend/main.cc: Upgraded Jsoncpp API
- OAuth2Backend/test/test_main.cc: Upgraded Jsoncpp API
- OAuth2Backend/test/EnvConfigTest.cc: Upgraded Jsoncpp API

Tested on Windows MSVC 2022, verified cross-platform compatibility for Linux/macOS.

+++ OAuth2Backend/plugins/OAuth2Plugin.h
@@ -84,9 +84,9 @@ class OAuth2Plugin : public drogon::Plugin<OAuth2Plugin>
     std::string storageType_;
 
     // TTL Configuration (Seconds)
-    long authCodeTtl_{600};
-    long accessTokenTtl_{3600};
-    long refreshTokenTtl_{3600 * 24 * 30};
+    long long authCodeTtl_{600};
+    long long accessTokenTtl_{3600};
+    long long refreshTokenTtl_{3600 * 24 * 30};
 
     void initStorage(const Json::Value &config);
 };
